"Close bug" fix and capture "senão" structure
Improved:

- "senão" is also detected when typed, making it no longer necessary to "backtab".

Fixed:

- Fixed titleBar widget rules (it no longer resizes if closeEvent is fired).
- Keyboard shortcuts were not captured on close event.

+++ src/pif_ide/pifeditor/pifeditor.cpp
@@ -2,8 +2,8 @@
 
 PIFEditor::PIFEditor(QWidget *parent) : QTextEdit(parent),
     prevTabulation("^(\\t*)\\.*"),
-    hasNewBlockE("^\\t*\\b(se|senão se|enquanto|faça|programa)\\b(.*)"),
-    hasNewBlock("^\\t*\\b(se|senão se|enquanto|faça|programa)\\b\\s(.*)")
+    hasNewBlockE("^\\t*\\b(se|senão se|senão|enquanto|faça|programa)\\b(.*)"),
+    hasNewBlock("^\\t*\\b(se|senão se|senão|enquanto|faça|programa)\\b\\s(.*)")
 {
 
 }
@@ -19,9 +19,9 @@ void PIFEditor::keyPressEvent(QKeyEvent *event){
         QTextCursor textCursor = this->textCursor();
         textCursor.beginEditBlock();
         QRegularExpressionMatch match = hasNewBlockE.match(this->document()->findBlockByNumber(textCursor.blockNumber() - 1).text());
-        if (match.hasMatch() && match.captured(2).isEmpty()) {
+        if (match.hasMatch() && match.captured(2).isEmpty() && this->textCursor().block().text().isEmpty()) {
             QString structure = match.captured(1);
-            QRegularExpressionMatch matchTabs = prevTabulation.match(this->document()->findBlockByNumber(this->textCursor().blockNumber()).text());
+            QRegularExpressionMatch matchTabs = prevTabulation.match(this->document()->findBlockByNumber(this->textCursor().blockNumber() - 1).text());
             if (structure == "faça"){
                 textCursor.deletePreviousChar();
                 textCursor.insertText(QString(" \n"
@@ -33,7 +33,7 @@ void PIFEditor::keyPressEvent(QKeyEvent *event){
                 textCursor.deletePreviousChar();
                 textCursor.insertText(QString(" NOME_DO_PROGRAMA\n"
                                               "%1\t\n"
-                                              "fim-programa").arg(matchTabs.hasMatch() ? matchTabs.captured(1) : ""));
+                                              "%1fim-programa").arg(matchTabs.hasMatch() ? matchTabs.captured(1) : ""));
                 textCursor.movePosition(QTextCursor::Up);
                 textCursor.movePosition(QTextCursor::EndOfLine);
             }
@@ -52,6 +52,16 @@ void PIFEditor::keyPressEvent(QKeyEvent *event){
                                               "%1fim-enquanto").arg(matchTabs.hasMatch() ? matchTabs.captured(1) : ""));
                 textCursor.movePosition(QTextCursor::Up);
                 textCursor.movePosition(QTextCursor::EndOfLine);
+            }
+            else if (structure == "senão"){
+                textCursor.insertText(matchTabs.hasMatch() ? matchTabs.captured(1) : "");
+                if (matchTabs.hasMatch() && matchTabs.captured(1).split("\t").size()){
+                    textCursor.movePosition(QTextCursor::Up);
+                    textCursor.movePosition(QTextCursor::StartOfLine);
+                    textCursor.deleteChar();
+                    textCursor.movePosition(QTextCursor::Down);
+                    textCursor.movePosition(QTextCursor::EndOfLine);
+                }
             }
             else {
                 textCursor.insertText("\t");
@@ -82,7 +92,7 @@ void PIFEditor::keyPressEvent(QKeyEvent *event){
             if (structure == "programa"){
                 textCursor.insertText(QString("NOME_DO_PROGRAMA\n"
                                               "%1\t\n"
-                                              "fim-programa").arg(matchTabs.hasMatch() ? matchTabs.captured(1) : ""));
+                                              "%1fim-programa").arg(matchTabs.hasMatch() ? matchTabs.captured(1) : ""));
                 textCursor.movePosition(QTextCursor::Up, QTextCursor::MoveAnchor, 2);
                 textCursor.movePosition(QTextCursor::EndOfLine);
                 textCursor.select(QTextCursor::WordUnderCursor);
@@ -109,6 +119,18 @@ void PIFEditor::keyPressEvent(QKeyEvent *event){
                 textCursor.insertText(QString("\n"
                                               "%1\t\n"
                                               "%1enquanto CONDIÇÃO").arg(matchTabs.hasMatch() ? matchTabs.captured(1) : ""));
+                textCursor.select(QTextCursor::WordUnderCursor);
+            }
+            else if (structure == "senão") {
+                textCursor.insertText(QString("se CONDIÇÃO então\n"
+                                              "%1").arg(matchTabs.hasMatch() ? matchTabs.captured(1) : ""));
+                textCursor.movePosition(QTextCursor::Up);
+                if (matchTabs.hasMatch() && matchTabs.captured(1).size()){
+                    textCursor.movePosition(QTextCursor::StartOfLine);
+                    textCursor.deleteChar();
+                }
+                textCursor.movePosition(QTextCursor::EndOfLine);
+                textCursor.movePosition(QTextCursor::WordLeft, QTextCursor::MoveAnchor, 2);
                 textCursor.select(QTextCursor::WordUnderCursor);
             }
             textCursor.endEditBlock();
